pause% for chord pattern; bugfix: shift generation limits, FULL pattern display fix, applypausechance leave at least 1 hit

+++ midimasterpiece/src/main/java/org/vibehistorian/vibecomposer/Enums/RhythmPattern.java
@@ -44,6 +44,6 @@ public enum RhythmPattern {
 		for (int i = 0; i < pattern.length; i++) {
 			counter += pattern[i];
 		}
-		return counter / pattern.length;
+		return counter / (double) pattern.length;
 	}
 }
